entrega-final

+++ src/main/java/dds/monedero/model/Movimiento.java
@@ -1,21 +1,22 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Movimiento {
   private LocalDate fecha;
   //En ningún lenguaje de programación usen jamás doubles para modelar dinero en el mundo real
   //siempre usen numeros de precision arbitraria, como BigDecimal en Java y similares
-  private double monto;
+  private BigDecimal monto;
   private TipoMovimiento tipo;
 
-  public Movimiento(LocalDate fecha, double monto, TipoMovimiento esDeposito) {
+  public Movimiento(LocalDate fecha, BigDecimal monto, TipoMovimiento esDeposito) {
     this.fecha = fecha;
     this.monto = monto;
     this.tipo = esDeposito;
   }
 
-  public double getMonto() {
+  public BigDecimal getMonto() {
     return monto;
   }
