+++ src/main/java/dds/monedero/model/TipoMovimiento.java
@@ -1,8 +1,10 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
+
 public interface TipoMovimiento {
 
   public abstract boolean isDeposito();
 
-  public abstract double calcularSaldo(double saldo, double monto);
+  public abstract BigDecimal calcularSaldo(BigDecimal saldo, BigDecimal monto);
 }
